% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMSE.R
\name{deriv_crit_IMSE}
\alias{deriv_crit_IMSE}
\title{Derivative of crit_IMSE}
\usage{
deriv_crit_IMSE(x, model, Wijs = NULL)
}
\arguments{
\item{x}{matrix for the new design (size 1 x d)}

\item{model}{\code{homGP} or \code{hetGP} model}

\item{Wijs}{optional previously computed matrix of Wijs, see \code{\link[hetGP]{Wij}}}
}
\value{
Derivative of the sequential IMSPE with respect to \code{x}
}
\description{
Derivative of crit_IMSE
}
\seealso{
\code{\link[hetGP]{crit_IMSE}} for the criterion
}
