% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.R
\name{deriv_crit_logEI}
\alias{deriv_crit_logEI}
\title{Derivative of logEI criterion for GP models}
\usage{
deriv_crit_logEI(x, model, cst = NULL, preds = NULL)
}
\arguments{
\item{x}{matrix for the new design (size 1 x d)}

\item{model}{\code{homGP} or \code{hetGP} model}

\item{cst}{threshold for contour criteria}

\item{preds}{pre-computed preds for contour criteria}
}
\description{
Derivative of logEI criterion for GP models
}
\references{
Ginsbourger, D. Multiples metamodeles pour l'approximation et l'optimisation de fonctions numeriques multivariables Ecole Nationale Superieure des Mines de Saint-Etienne, Ecole Nationale Superieure des Mines de Saint-Etienne, 2009 \cr \cr
Roustant, O., Ginsbourger, D., DiceKriging, DiceOptim: Two R packages for the analysis of computer experiments by kriging-based metamodeling and optimization, Journal of Statistical Software, 2012
Ament, S., Daulton, S., Eriksson, D., Balandat, M., & Bakshy, E. (2024). Unexpected improvements to expected improvement for Bayesian optimization. Advances in Neural Information Processing Systems, 36.
}
\seealso{
\code{\link[hetGP]{crit_logEI}} for the criterion
}
