% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctions.R
\name{f1d}
\alias{f1d}
\title{1d test function (1)}
\usage{
f1d(x)
}
\arguments{
\item{x}{scalar or matrix (size n x 1) in [0,1]}
}
\description{
1d test function (1)
}
\examples{
plot(f1d)
}
\references{
A. Forrester, A. Sobester, A. Keane (2008), Engineering design via surrogate modelling: a practical guide, John Wiley & Sons
}
