% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TestFunctions.R
\name{f1d2}
\alias{f1d2}
\title{1d test function (2)}
\usage{
f1d2(x)
}
\arguments{
\item{x}{scalar or matrix (size n x 1) in [0,1]}
}
\description{
1d test function (2)
}
\examples{
plot(f1d2)
}
\references{
A. Boukouvalas, and D. Cornford (2009), Learning heteroscedastic Gaussian processes for complex datasets, Technical report. \cr \cr
M. Yuan, and  G. Wahba (2004), Doubly penalized likelihood estimator in heteroscedastic regression, Statistics and Probability Letters 69, 11-20.
}
