% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMSE.R
\name{IMSPE}
\alias{IMSPE}
\title{Integrated Mean Square Prediction Error}
\usage{
IMSPE(X, theta = NULL, Lambda = NULL, mult = NULL, covtype = NULL,
  nu = NULL, eps = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{\code{hetGP} or \code{homGP} model. Alternatively, one can provide a matrix of unique designs considered}

\item{theta}{lengthscales}

\item{Lambda}{diagonal matrix for the noise}

\item{mult}{number of replicates at each design}

\item{covtype}{either "Gaussian", "Matern3_2" or "Matern5_2"}

\item{nu}{variance parameter}

\item{eps}{numerical nugget}
}
\description{
IMSPE of a given design
}
\details{
One can provide directly a model of class \code{hetGP} or \code{homGP}, or provide \code{X} and all other arguments
}
