% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IMSE.R
\name{Wij}
\alias{Wij}
\title{Compute double integral of the covariance kernel over a [0,1]^d domain}
\usage{
Wij(mu1, mu2 = NULL, theta, type)
}
\arguments{
\item{mu1, mu2}{input locations considered}

\item{theta}{lengthscale hyperparameter of the kernel}

\item{type}{kernel type, one of "\code{Gaussian}", "\code{Matern5_2}" or "\code{Matern3_2}", see \code{\link[hetGP]{cov_gen}}}
}
\description{
Compute double integral of the covariance kernel over a [0,1]^d domain
}
\references{
M. Binois, J. Huang, R. Gramacy, M. Ludkovski (2017+), Replication or exploration? Sequential design for stochastic simulation experiments.
}
