% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hetGP.R
\name{compareGP}
\alias{compareGP}
\title{Likelihood-based comparison of models}
\usage{
compareGP(model1, model2)
}
\arguments{
\item{model1, model2}{\code{hetGP} or \code{homGP} models}
}
\value{
Best model based on the likelihood, first one in case of a tie
}
\description{
Compare two models based on the log-likelihood for \code{hetGP} and \code{homGP} models
}
\note{
If comparing homoskedastic and heteroskedastic models, the un-penalised likelihood is used for the later, see e.g., (Binois et al. 2017+).
}
\references{
M. Binois, Robert B. Gramacy, M. Ludkovski (2017+), Practical heteroskedastic Gaussian process modeling for large simulation experiments, arXiv preprint arXiv:1611.05902.
}
